Introduce a feature flag to turn the compilation of Python bindings on and off.
