+++ Cargo.toml
@@ -15,6 +15,7 @@ crate-type = ["cdylib", "rlib"]
 # you don't need it as long as you have Z3 installed (because your
 # build may take >30min if you enable it).
 static-z3 = ["z3/static-link-z3"]
+algorithms_pyo3_bindings = []  # Feature flag for reachability analysis
 
 [dependencies]
 pyo3 = { version = "0.23.5", features = ["abi3-py37", "extension-module", "num-bigint", "py-clone"] }
