+++ src/bindings/algorithms/cancellation/tokens.rs
@@ -8,7 +8,6 @@ use std::{
 };
 
 use log::info;
-use pyo3::Python;
 
 use crate::bindings::algorithms::cancellation::CancellationHandler;
 
@@ -104,32 +103,3 @@ impl CancelTokenTimer {
 }
 
 /* Timer - End */
-
-/* Python - Start */
-
-/// A [CancellationHandler] that wraps any ohter [CancellationHandler] and also checks for Python
-/// interrupts.
-#[derive(Clone, Debug, Default)]
-pub struct CancelTokenPython(Box<dyn CancellationHandler>);
-
-impl CancellationHandler for CancelTokenPython {
-    fn is_cancelled(&self) -> bool {
-        if self.0.is_cancelled() {
-            return true;
-        }
-
-        Python::with_gil(|py| py.check_signals()).is_err()
-    }
-
-    fn start_timer(&self) {
-        self.0.start_timer()
-    }
-}
-
-impl CancelTokenPython {
-    pub fn with_inner<T: CancellationHandler + 'static>(handler: T) -> Self {
-        CancelTokenPython(Box::new(handler))
-    }
-}
-
-/* Python - End */
