+++ src/bindings/algorithms/fixed_points/mod_python/_impl_pyerr.rs
