+++ src/bindings/algorithms/fixed_points/mod_python/mod.rs
@@ -1,0 +1,20 @@
+use pyo3::{
+    types::{PyModule, PyModuleMethods as _},
+    Bound, PyResult,
+};
+
+use super::{FixedPoints, FixedPointsConfig, FixedPointsError};
+
+mod _impl_pyerr;
+mod fixed_points_config_python;
+mod fixed_points_impl_python;
+
+pub use fixed_points_config_python::PyFixedPointsConfig;
+use fixed_points_impl_python::PyFixedPoints;
+
+pub fn register(module: &Bound<'_, PyModule>) -> PyResult<()> {
+    module.add_class::<PyFixedPoints>()?;
+    module.add_class::<PyFixedPointsConfig>()?;
+
+    Ok(())
+}
