+++ src/bindings/algorithms/percolation/mod_python/_impl_pyerr.rs
@@ -1,10 +1,8 @@
 use pyo3::PyErr;
 
 use crate::bindings::algorithms::{
-    cancellation::CancelledError, configurable::CreationFailedError,
+    cancellation::CancelledError, configurable::CreationFailedError, percolation::PercolationError,
 };
-
-use super::PercolationError;
 
 impl From<PercolationError> for PyErr {
     fn from(err: PercolationError) -> Self {
