+++ src/bindings/algorithms/percolation/mod_python/percolation_config_python.rs
@@ -3,7 +3,7 @@ use std::time::Duration;
 use pyo3::{pymethods, PyResult};
 
 use crate::bindings::algorithms::{
-    cancellation::tokens::{CancelTokenPython, CancelTokenTimer},
+    cancellation::{CancelTokenPython, CancelTokenTimer},
     configurable::Config as _,
     graph_representation::PyGraphRepresentation,
 };
