+++ src/bindings/algorithms/percolation/mod_python/percolation_impl_python.rs
