+++ src/bindings/algorithms/percolation/mod_python/subspace_representation.rs
