+++ src/bindings/algorithms/trap_spaces/mod_python/_impl_pyerr.rs
