+++ src/bindings/algorithms/trap_spaces/mod_python/trap_spaces_impl_python.rs
